+++ apps/intellij/src/main/kotlin/dev/nx/console/generate/ui/editor/NxGenerateUiFile.kt
@@ -8,13 +8,13 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.testFramework.LightVirtualFile
 import com.intellij.ui.jcef.*
 import com.intellij.util.ui.UIUtil
-import dev.nx.console.NxConsoleSettingsProvider
 import dev.nx.console.NxIcons
 import dev.nx.console.generate.run_generator.RunGeneratorManager
 import dev.nx.console.generate.ui.CustomSchemeHandlerFactory
 import dev.nx.console.generate.ui.utils.getHexColor
 import dev.nx.console.generate.ui.utils.onBrowserLoadEnd
 import dev.nx.console.models.NxGenerator
+import dev.nx.console.settings.NxConsoleSettingsProvider
 import javax.swing.Icon
 import javax.swing.JComponent
 import kotlinx.serialization.decodeFromString
