+++ apps/intellij/src/main/kotlin/dev/nx/console/settings/NxConsoleSettingsConfigurable.kt
@@ -1,4 +1,4 @@
-package dev.nx.console
+package dev.nx.console.settings
 
 import com.intellij.openapi.options.SearchableConfigurable
 import com.intellij.ui.dsl.builder.panel
